requirements.txt
gradio
gradio-pdf
langchain-core
PyMuPDF
azure-storage-blob